MessagesFromMCU ArduinoV1

+++ NSUWatcher/CommandCenter/MessagesFromMcu/Factory/ArduinoV1/Data/WaterBoilerSnapshot.cs
@@ -29,10 +29,16 @@ namespace NSUWatcher.CommandCenter.MessagesFromMcu.Factory.ArduinoV1.Data
         public bool ElHeatingEnabled { get; set; }
         
         [JsonProperty(JKeys.WaterBoiler.ElPowerChannel)]
-        public byte ElPowerChannel { get; set; }
+        public int ElPowerChannel { get; set; }
         
         [JsonProperty(JKeys.WaterBoiler.PowerData)]
         public IElHeatingDataSnapshot[] ElHeatingData { get; set; } = new HeatingData[0];
+        
+        //[JsonProperty(JKeys.Generic.Source)]
+        //public string Source { get; set; } = string.Empty;
+        
+        //[JsonProperty(JKeys.Generic.CommandID)]
+        //public string? CommandID { get; set; }
     }
 #nullable disable
